fix:秋丸　ログオフメソッドをstartMenuコントローラからログインコントローラへ移動 帳票出力機能　ユーザー検索機能で出力ボタンの活性非活性連動を正しく修正、勤怠帳票、勤怠日報帳票作成のため編集(HTML未修整のためエラー箇所あり) 勤怠管理機能(社員権限)　勤怠表に縦スクロール追加　ヘッダー、勤怠状況optionタグを簡単な方法に修正

+++ nippo_kintai_project/src/main/java/com/analix/project/controller/LoginController.java
@@ -1,11 +1,13 @@
 package com.analix.project.controller;
 
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 
 import com.analix.project.entity.WorkSchedule;
 import com.analix.project.service.LoginService;
@@ -13,19 +15,18 @@ import com.analix.project.service.WorkScheduleCacheService;
 import com.analix.project.service.WorkScheduleService;
 import com.analix.project.util.SessionHelper;
 
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
 import jakarta.servlet.http.HttpSession;
-
+import lombok.RequiredArgsConstructor;
+@RequiredArgsConstructor
 @Controller
 public class LoginController {
 
-	@Autowired
-	private LoginService loginService;
-	@Autowired
-	private SessionHelper sessionHelper;
-	@Autowired
-	private WorkScheduleService workScheduleService;
-	@Autowired
-	private WorkScheduleCacheService workScheduleCacheService;
+	private final LoginService loginService;
+	private final SessionHelper sessionHelper;
+	private final WorkScheduleService workScheduleService;
+	private final WorkScheduleCacheService workScheduleCacheService;
 
 	@GetMapping("/")
 	public String getLogin(Model model, HttpSession session) {
@@ -37,13 +38,13 @@ public class LoginController {
 			HttpSession session, Model model) {
 		System.out.println(sessionHelper.getUser());
 		String loginResult = loginService.handleLogin(employeeCode, password);
-		
+
 		// 勤務時間をキャッシュに追加
 		WorkSchedule workSchedule = workScheduleService.getWorkScheduleEntity(sessionHelper.getUser().getId());
 		if (workSchedule != null) {
 			workScheduleCacheService.put(sessionHelper.getUser().getId(), workSchedule);
 		}
-		
+
 		if (loginResult.equals("error")) {
 			model.addAttribute("error", "ユーザーID、パスワードが不正、もしくはユーザーが無効です。");
 			return "common/login";
@@ -60,19 +61,37 @@ public class LoginController {
 		return "redirect:/common/startMenu";
 	}
 
+	/**
+	 * 『ログオフ』ボタン押下後
+	 * @return ログイン画面遷移
+	 */
+	@RequestMapping(path = "/common/logoff", method = RequestMethod.POST)
+	public String logoff(HttpServletRequest request, HttpServletResponse response,
+			RedirectAttributes redirectAttributes) {
+		if (sessionHelper != null) {
+			workScheduleCacheService.remove(sessionHelper.getUser().getId()); // キャッシュから勤務時間削除
+		}
+		HttpSession session = request.getSession(false);
+		if (session != null) {
+			session.invalidate();
+		}
+		redirectAttributes.addFlashAttribute("message", "ログオフしました。再度ログインしてください。");
+
+		return "redirect:/";
+	}
+
+	/**
+	 * タイムアウト時
+	 * @param model
+	 * @param session
+	 * @return ログイン画面遷移
+	 */
 	@GetMapping("/timeout")
 	public String timeout(Model model, HttpSession session) {
-		session.invalidate(); // セッションを無効化
-		workScheduleCacheService.remove(sessionHelper.getUser().getId()); // キャッシュから勤務時間削除
 
+		session.invalidate(); // セッションを無効化
 		model.addAttribute("message", "タイムアウトしました。再ログインしてください。");
 		return "common/login";
 	}
 
-	//	// ログアウト処理
-	//	@GetMapping("/logout")
-	//	public String logout(HttpSession session) {
-	//		session.invalidate(); // セッションを無効化
-	//		return "redirect:/login"; // ログインページへリダイレクト
-	//	}
 }
